Incorporation of the CLaMS model data input into the MSS environment. The files mslib/mss_config.py.sample msui/mss_settings.py.sample mswms/mss_wms_settings.py.sample contain changes that need to be completed by local path data and would need to be copied to the correponding file names without ".sample"
The CLaMS data should be placed as cf-conform NetCDF file into the directory speciefied in
mslib/mss_config.py
in the dictonary "datapath"

In the user interface, a map section for India is added.
The setup of CLaMS species corresponds to the setup needed for the Stratoclim campaign
based in India. It contains chemical species from the "simplified Chemistry verstion"
(Pommrich et al., Geophs. Model Dev. 7, 2895-2916, 2014) and surface origin tracers
(Vogel et al., Atmos. Chem. Phys., 15, 13699-13716, 2015).

The array of chemical species and tracers is defined consistently at different parts of the
programm
(
  * data_organisation table in class CLAMSDataAccess
  * define the Netcdf variable definitions used in CFVariableIdentifier
  * default value range definition in horizontal and vertical plots
  * creation of the classes "HS_ChemStyle_PL_xxx" and "VS_ChemStyle_PL_xxx"
)