Fixes issue #73
Strangely, the title text was updated correctly in Matplotlib, and also
commanded to be redrawn, but only following a system-event like a mouse
click opening another widget or a switch to another application, the
text was actually updated. Adding a QT repaint command solves the issue.
Maybe a bug in QT.