refactored remote sensing functionail to be controlled using a docking widget instead of map appearance dialog.
first step in the refactoring, which moves the user interface to a less
exposed position. This also allows further configurability with respect
to, e.g., colour scale of solar angle and viewing direction with respect
to flight direction.

For simplicities sake, the barely working addition of tangent point
height to plot title was removed as no informations is currently deemed
better than wrong information, which could easily happen for the oild
implementation if in addition to the TP altitude also layers or levels
were modified.

Further steps should also move more functionality from the
mpl-path-interactor to the remote sensing docking widget.