Provided a simplified way of calculating and managing performance data. Partially adresses issue #85
This patch completely restructures the calculation of performance
data. The old way using a client-server infrastructure is replaced by a
very simple on-the-fly one that, by necessity, does not take temperature
and wind speed along the flight path into account. The data format is
also changed to a JSON based one, which can be dynamically loaded using
the GUI. The current aircraft is saved and restored when the GUI is
restarted.