Reworked solar angle in RS dock widget to use UTC time from waypoint table. Related to issue #15.
With this update, the UTC times from waypoint table will be used to
determine the solar angle instead of the very rough approximation of
perforamnce calculation used before.

On the other hand, the solar angle is not correctly represented while a
waypoint is moved with the mouse as the Table is only updated when
releasing the button. This could be fixed by updating the table also
when moving the mouse at quite a performance cost.