Adds support for more features of KML.  Fixes issue #92. Related to issue #61.
The KML Overlay now also supports Points and single Polygons as well as
MultiGeometries of both types and Folders. LinearStrings are unsupported
as well as all kinds of styles. Style support in some simple fashion
should be added as well as zordering of overlay? Or support for defining
draw style in the GUI.