Rework of Text plugins and exceptions handling.
Discussions with DLR-FX have shown that FliteStar indeed has no import
function. As such the flitestar plugin has been split into the read
routine for the FliteStar Text output format and the seemingly ad-hoc
Textformat created for the POLSTRACC campaign.

In addition, SyntaxErrors are now caught by the GUI to present errors in
read files to the USER outside the logging.