Show QMesagebox in case of syntax error in JSON file
A syntax error in the configuration JSON file will throw a ValueError
(shouldn't that be a SyntaxError?!), which will dump the stack and show
the error message. At least using windows, the cmd window containing the
log will be immediately closed preventing any sensible debugging. Now,
in addition to the stack trace etc. a QMessageBox will be opened with
some information about the error. I think that an error in the
configuration file is rightly handled by an abort of the application.