Catch errors in plugin import/export handling gracefully.
Errors in the configuration of io plugins will now be caught and show a
QMessageBox. The affected plugin will, obviously, not be loaded, but the
application will continue working without the problematic pluging.
As basically everything can go wrong with the import of the
user-supplied module, a wild-card exception catch is required around
the import statement.