Use a QMessagebox to supply information about MSSUI crashing Exceptions
As the console may not be visible, a GUI feedback about errors is
required. The solution presented consistently uses the logging framework
(for output in logfiles), the console, and a QMessagebox to dump
information about the error. This should allow easier identification of
errors on Windows and potentially other systems, where MSSUI is started
using a GUI.