Added support for default "EPSG" projections
It is a hassle that EPSG projections need to be manually and
consistently configured in MSUI and MSWMS server. To simplify especially
the stereographic projections, I added a feature that the user-defined
EPSG codes starting with 777 and 778 are automatically recognized as
stereographic projections with projections centers given as llLLL
following the starting codes like suggested in the examples. The given
infrastructure can be readily extended by more MSS specific or generally
valid codes.

The MSWMS server still needs to specify all EPSG codes that he wants to
announce that he supports, however the configuration may be simplified
to simply list the code and a "None" and the parameters will be
automatically filled.