Draft of simplifying the generation of beautiful plots
The generic class is able to plot about everything, but is currently
very limited with respect to customization. This draft adds some
configuration options to the Generic plotting class that allow it to
supplant the dedicated classes for GW, N^2 and tropopause altitude.
In addition some existing generic plots of e.g. PV is also enhanced by a
dedicated colourmap.

This draft only fully realizes the update for horizontal plots as
vertical ones are a tad more complicated due to the vertical axes being
different.