Update to generic vertical styles to make them much more configurable
This update also adds a lot of funtions to the make_generic function in
vertical style section similar to the function offered in the horizontal
plot section. This also allows the definition of various contours and
plotting styles. Due to the conversion of tropopausealtitude from
GPS altitude to pressure altitude the current GW plots cannot make use
of the new class. The new dataformat will do the conversion outside of
MSS and thus allow the removal of at least two classes.