Rework of threading in WMS fetching to QThread
Refactoring of current background WMS-MAP and legend fetching to the use
of QThreads instead of threading and Queue library. This simplifies the
code a bit and uses the more Qt-y event/signal parsing and does not use
a while loop to wait for server to return data. Due to the nature of now
three threads and their intercommunication, much can go wrong, but as
the legend retrieval was integrated into the WMS map retrieval, it is in
effect simpler than before, even though the progress dioalog is a bit
difficult to close at the correct time under all (erroneous)
circumstances.