Added basic support of Styles defined in KML file
Parse linewidth and colour properties of KML file. Neither StyleMaps nor
external style definitions are supported, yet.

Added a facility to define linewidth of KML overlays and added a button
to overwrite the behaviour defined by the styles, if desired. User
defined colour and linewidths are used for KML features that do not
reference a style.