Added more info to service and a human-readable view on capabilities
The standard offers more information that may be supplied in the
capabilities document, especially the email field for getting in
contact.

Also the view of the capabilties was more useful to the developer than
the user and has been enhanced by a checkBox that toggles between a
human-readable view on the general information of the server (title,
abstract, contact, etc.) but hides layers and a view on the raw xml
document.