Added tooltips to many widgets and redesigned about dialogue.
Tooltips were only used in WMS control dialog. Now all widgets that are
not fully self-documenting use tooltips for further explanation. More
work could be employed here.

The About dialog required scrolling to read the Text. It has been
redesigned to give more space, if available to the TextBox. Also a link
to the issue tracker was added and both links were made "clickable" to
open the URLS in an external browser.