Fixed display and removal of legend graphics
Using http://osmwms.itc-halle.de/maps/osmfree?SERVICE=WMS as a test
server, the legend display feature was tested. There were some issues
with delayed-display and the legend not being removed when displaying
other graphics later on. Also canceling the retrieval of the image would
not cancel the display if a legend was retrieved later on due to the
concurrency of the processes.