Add a facility to better check for a redirection of WMSURL
If the service uses a redirection for some reason or simply a very
different URL for the Get and Post requests, the wms control will not
execute the map request. The check has been reduced to see that the
current text box has still the same content as upon getcapabilities
call. in addition the getmap/view buttons will be disabled if the
combobox content is not consistent with the current WMS object URL.