Further enhanced WMS control by service cache and intuitive buttons
Now the WMS control remembers the URLs it requested XML capabilities for
and restores the buttons for the known URL upon entering it. Also *all*
controls will be disabled when changing the URL (and reenabled when
correcting it).
This enforces consistency between request WMS maps and URL Combobox. It
also allows for the simple switching between multiple servers and would
allow the split of layers upon different "sub"servers to reduce the
amount of layers in the layer combo box.

Each WMS server needs to be requested manually, but one might preload
XML documents for preconfigured servers in the background.