Add facility to add derived data products to generic plots
There is a method for providing additional data products based on model
data. This method is now used to provide
air_pressure/potential_temperature_in_air data for vertical sections.
Also, a parameter to make_generic was added to override this method with
a suer defined function to provide additional data for contour/contourf
plots.