Catch ValueError when opening config-pickles
When opening a Python2 pcikle file with Python3 (and maybe vice-versa),
a ValueError may be emitted due to the different string handling.
A full solution should ideally enable the sharing of pickle files
between different Python and (to some extent) MSS versions.
A crash of the application is certainly a no go.

Resolves issue # 198