Check for existance of setting pickle file for better logging
Previously, a new MSS installation would dump warnings to the console,
if no settings file were present, even though that is expected. The new
behaviour now checks for the presence and does not warn in case no file
is present. If the existing file cannot be loaded, however, this is now
logged as error. This error is not alerted to the user via the GUI,
though.