Enable the use of fractional elevation in horizontal layers
In some coordinate systems, fractional elevations may be useful. Added
support by WMS server for fractional layers. Added demodata and
associated testcase using fractional layer.

Changed behaviour of server to raise a service exception, if a
horizontal elevation is requested, for which no data is available. That
is according to the WMS standard. Future work may implement the use of
the neasrest neighbour, but this must be noticed in the capability xml
and noted by a http warning.

Resolve isse #203