Added a new, improved data access class
The new data access class does not require the data files to be named
according to a scheme, but uses, aside a string identifier, solely
information from within the files to build up a data structure for
accessing data.

The init time is determined from the reference time "0". The valid times
are determined by the time values. Variables are collected according to
the standard_names. File type (pl/ml/etc) is determined by the available
coordinate axis.

As before, each coordinate type must be placed within its own file.
However the class might be extended to relax this requirement, if
desired.

The class was tested and seems to be able to replace ECMWF and CLaMS
access classes, at least for the files provided by demodata.