Add an exception to catch strange error in moving points in TopView
Under unknown circumstances that could not be reproduced, the Table was
updated by point moving with an object that generated a "TypeError" when
passing it to a "float()" function.
To prevent a crash from MSS, that error is now captured at the
appropriate places and an error is dumped to the log, which might
help in finding the root cause.

Fix issue #238