Make wms_control cache servicing more resilient
When a file cannot be deleted (access rights?) the MSS application was killed. The new update makes it more resilient by catching Exceptions and notifying hte user both in the log and with a pop-up window about what occured. To not cause hundreds of error messages, the servicing is aborted upon the first error. That shouldn't cause any problems in general.

Fix issue #243