Make pyqt4 behave more like pyqt5 in QVariant conversion
pyqt5 is able to retrieve the original type from a QVariant object.
In pyqt4 one has to manually retrieve it. The current implementation
just converts it to a string, which is afterwards converted to a number.
In case of moving a point with the mouse, the float was first converted
to a QVariant and then to a double using the toString method.
This conversion however delivered a US formatted number as string, which
afterwards caused problems when converting it back to a float using
localization.

The current code prevents this reconversion by accessing directly the
float.

Fix issue #255