Added a menu entry for activating flight track
As double-clicking may not be "obvious", a menu entry should help the
novice to switch between flight tracks.
Some refactoring with last remembered directory. The last directory used
for saving FTML files is now used as default for all file operations.
Might be generalized and or remembered between sessions..?

See #217