Add a very simple unit conversion routine for the most obvious problems.
The original DLR models use mostly Pa for pressure and m^2s^-2 for
geopotential altitude, while FZJ ones use often hPa and m or km.
Instead of assuming the model to be in a given unit, the units attribute
can be read out. This has already been implemented in a past patch, but
was not consistently used so far. There are also libraries for unit
conversion, but they fail, e.g., when converting GPH in m^2/s^2 to m.
The current solution is very simple, extendable, and effective
replaceable by a more professional solution by putting all conversion in
one place.

Fix issue #328