Remove check if file is part of available_files()
In previous data access classes, filenames could be returned that were
non-existant. Thus a check existet for this.

However, the current implementations do not do that. Instead, they
identify new files on the fly. The current implementation does not take
that into account as it compares against a list of known files *before*
the potential update.

The fix removes the check, as the ValueError will be raised by the
DataAccess class in case no file is available.

The question is only, if we need to check at this point in time, if a
file has "vanished" from the disk due to being deleted.

Fix issue #425