Made generating available elevation levels more efficient
To generate available elevation levels the plot object needs
to identify one timestep, for which data is available.
For certain (stupid, but legal) combinations of init_times, this
triggered multiple re-reads of available data.
However, the data access class is always updated before this method
is called. The code was simplified and restructured in a way that
the data access class does not re-read files from disk by adding a
have_data method, which determines *known* data.

Fix issue #432