Fixed timing issue in zooming/redraw
Due to the event handling, it was previously random if the zoom or the
redraw was triggered first. Now, the redraw occurs always after the
zoom/pan event.

The issue is not closed yet, as further testing is required if there are
other error sources. If yes, the REDRAW button may be removed.

See issue #437