Fixed time handling for skyfield
In contrast to most other packages, skyfield requires a timezone in its
datetime objects. We are using UTC already, but without stating it (as
tz info causes a lot of problems). We follow skyfield guidelines by
replacing the timezone instead of calling localize, which causes
issues in the current version.