Some tidying up of handling imports in MSCOlab
The connect seems to be wrong, a disconnect seems in order.
It might be better to not create a new waypoint_model, but change the
content of the old one?
Either way, with these streamlining, the error was not reproducible
anymore. Since it was sporadic before, this might be by chance.

See #1960