Fix test_add_users_with{,out}_updating_credentials_in_config_file
The overly aggressive mock as a function decorator hid the mscolab
server update dialog from create_msui_settings_file as well. The mock
should instead be applied at the innermost point that is possible. The
create_msui_settings_file is also replaced with a modify_config_file
call as to not overwrite the mscolab URL set in the mscolab_server
fixture.