Make mscolab_server a session-scoped fixture
The eventlet wsgi server used (now manually, before implicitly through
start_server) does not play nice with multiprocess.Process. I could not
find a way to make this a working function-scoped fixture. Having the
mscolab_server be a session-scoped fixture should make the test suite
faster though and so far it does look like it is sufficiently stateless
for the tests, i.e. the database resets are enough to avoid leakage
between tests.