Refactor Qt related tests to use pytest-qt
Using pytest-qt the QApplication instance can just be taken from its
qapp fixture instead of being setup in each test. The fixture to fail if
there are message boxes left at the end of a test can then be used by
making the qapp fixture request it, which means that only tests that
actually use Qt will have this fixture run (as opposed to the previous
autouse fixture, which always ran).