Add per-operation locks to FileManager (#2147)
Without these locks there is a race condition when writing to the
operations "main.ftml" file, which happens in multiple different
methods, e.g. save_file.  When the mscolab application is running in a
multi-threaded wsgi server it can happen that two threads try to write
to the same file, corrupting it in the process.