Give Qt some time to handle its events
In these tests in test_mscolab_admin_window.py it is possible that a
late signal triggers a call to render_new_permission, at which point the
mscolab_server_url is unset, which in turn leads to a TypeError in the
tests teardown. Adding a qWait at the ends of these tests should reduce
the likelihood of this issue, but won't eliminate it completely. I have
no idea what to wait for instead to fix this properly though.