Remove unnecessary development dependencies (#2140)
* Remove unnecessary development dependencies

We have a GitHub Action that runs flake8, which in turn wraps PyFlakes,
pycodestyle and McCabe. Since pep8 was renamed to pycodestyle those two
are the same. Adding flake8 to the dependencies therefore means we can
remove pep8 and pycodestyle. The pytest-pep8, which is really old, and
pytest-flake8 plugins also become redundant since we have the CI action
anyway. The pytest-cache plugin is redundant since it was integrated
into pytest in 2015 (pytest v2.8.0).

* Update development docs to use flake8 directly