Fix ConversionError with pint=0.23 (#2168)
The error happens when trying to use a pint Quantity in a matplotlib
axis that has no units set, because the attempted conversion to
`units=None` fails. This change sets the units of the axis to be the
same as the ticks' units, but only if the ticks have units set and the
axis does not. This is effectively a fallback to the Quantity's units if
nothing else is set.