Consolidate skips due to start method "fork" (#2166)
There is just one place where the start method "fork" is used, which is
in tests.fixtures. By adding a conditional skip at that location all
other skips for this reason become obsolete, since the tests will be
automatically skipped if they request one of the fixtures that require
this start method from multiprocessing and the start method is
unsupported on the given system.