Add xvfb-run to pytest execution
The testing-stable container image no longer includes pyvirtualdisplay
due to a change that made it unnecessary for develop that also affected
the stable image. Instead of reinstalling pyvirtualdisplay this just
wraps the entire pytest execution in xvfb-run, which has mostly the same
effect. This is a bandaid to fix the stable CI until all of the recent
changes to the tests are backported from develop to stable.