Install pyvirtualdisplay in testing workflow
The testing-stable container image no longer includes pyvirtualdisplay
due to a change that made it unnecessary for develop that also affected
the stable image. This installs the package as part of the testing
workflow instead. This is a band-aid to fix the stable CI until all of
the recent changes to the tests are backported from develop to stable.