Partially fix non-deterministic coverage (#2318)
* Partially fix non-deterministic coverage

This is a band-aid fix that mostly resolves the issue with
non-deterministic coverage. Without the wait, tests abruptly end and
coverage can fluctuate. With the wait, Qt has time to handle its stuff
for a while longer, which stabilizes the coverage. The proper fix would
be to clean up everything Qt-related that is created in a test after
each test, but I couldn't get that to work yet.

There are also at least two more cases of non-deterministic coverage
that seem to be unrelated to Qt and which aren't fixed with this.

* Add docstring to qtbot overwrite