Use find_namespace_packages
This fixes warnings about sub-directories of mslib/ that were previously
not part of the packages because they don't contain __init__.py, but
that were still recognized and added as packages due to a deprecated
behavior of setuptools.

The change from exclude to include is due to the excludes not being
exhaustive. Multiple consecutive builds would add the entirety of build/
into the package with the excludes. Since we know that mslib and its
sub-packages are the only ones that should be distributed we can just be
explicit about that.