Fix inconsistent SocketsManager objects
On import of mslib.mscolab.server initialize_managers was called once,
hooking events up to handlers in a SocketsManager instance and
initializing the Flask app with the Flask-SocketIO extension. Then, in
handle_start, initialize_managers was called a second time, creating
another SocketsManager object, overwriting the instance saved with the
SocketIO object, and trying to initialize the Flask app with it again.

The last step does not work though, a Flask app can only be initialized
once with each extension (i.e. the sockio.init_app(app) call can only be
done once per Flask app object).

What this led to is that the Flask-SocketIO extension was using the
first SocketsManager object to handle incoming events, while the server
side was using the second SocketsManager object when it called methods
on sockio.sm.

This changes the startup process to do the initialization just once.