Use Flask-SocketIO test client
Previously test_sockets_manager.py communicated with a MSColab server
running in a different process. This made it impossible to assert
anything on the server state (e.g. on attributes of the SocketsManager
object) due to the process boundary.

Using a test client also simplifies the test setup quite a bit.