Use flask-migrate for database migrations (#2432)
With this flask-migrate is used on application setup to automatically
migrate the configured database to the latest revision shipped with
MSColab. This will make manual migrations in production environments
unnecessary in the future.

Also included is an upgrade path from databases which were previously
manually managed. This entails generating a new database and copying
over all data.