Replace setup.py with pyproject.toml
This eliminates the need for the future module by setting the mslib
version in pyproject.toml and populating __version__ from package
metadata. One limitation of pyproject.toml is that it is not possible to
define OS-specific entrypoints, so the msidp application now has a
runtime check to error out on Windows instead of just not existing
there.

The pyproject.toml file is the latest PEP standard for python project
metadata. In the future it could also accomodate additional
configuration that are currently set in setup.cfg and pytest.ini.